Changes to actor speed specifications.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@2350 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/SkeletonAnim.h
@@ -25,9 +25,9 @@ public:
 	CSkeletonAnimDef* m_AnimDef;
 	// speed at which this animation runs
 	float m_Speed;
-	// time(s) during the animation at which the interesting bit(s) happens (msec)
-	size_t m_ActionPos;
-	size_t m_ActionPos2;
+	// time(s) during the animation at which the interesting bit(s) happens (fractional)
+	float m_ActionPos;
+	float m_ActionPos2;
 	// object space bounds of the model when this animation is applied to it
 	CBound m_ObjectBounds;
 };
